fix(MicMute): Fix toggling mute button on Teams and Zoom
Corrects the menu item names used to mute/unmute audio in Zoom as they are case-sensitive by making a regex that is case insensitive. This resolves issues where the script fails to correctly toggle the mute state due to changes in application menu.

Also changes the way the Microsoft teams app is found due to Teams change in bundle id.